requirements.txt
google-api-python-client>=2.130.0
google-auth>=2.30.0
google-auth-httplib2>=0.2.0
google-genai>=0.5.0
requests>=2.32.0
PyYAML>=6.0
pydantic>=2.7.0
click>=8.1.0
structlog>=24.1.0
tenacity>=8.3.0
python-dotenv>=1.0.0
